+++ CanvasPlusPlayground/Features/Assignments/View/CourseAssignmentsView.swift
@@ -184,9 +184,6 @@ private struct AssignmentRow: View {
                     type: .assignment
                 )
             }
-            .environment(\.openURL, OpenURLAction { url in
-                    .systemAction
-            })
             .tag(assignment)
         } else {
             bodyContents
